refactor(orchestrator): extract task lifecycle workflows
