+++ services/local-service/internal/orchestrator/service.go
@@ -280,258 +280,6 @@ func (s *Service) RunEngine() *runengine.Engine {
 	return s.runEngine
 }
 
-// SubmitInput handles agent.input.submit.
-// It captures context, derives intent suggestions, and decides whether the task
-// waits for more input, asks for confirmation, or runs immediately.
-func (s *Service) SubmitInput(params map[string]any) (map[string]any, error) {
-	snapshot := s.context.Capture(params)
-	options := mapValue(params, "options")
-	confirmRequired := boolValue(options, "confirm_required", false)
-	if response, handled, resolvedSessionID, err := s.maybeContinueExistingTask(params, snapshot, nil, taskContinuationOptions{
-		ConfirmRequired:      confirmRequired,
-		ForceConfirmRequired: confirmRequired,
-	}); err != nil {
-		return nil, err
-	} else if handled {
-		return response, nil
-	} else if strings.TrimSpace(resolvedSessionID) != "" {
-		params = withResolvedSessionID(params, resolvedSessionID)
-	}
-	suggestion := s.intent.Suggest(snapshot, nil, confirmRequired)
-	suggestion = s.normalizeSuggestedIntentForAvailability(snapshot, suggestion, confirmRequired)
-	if handledResponse, handled, err := s.handleScreenAnalyzeSuggestion(params, snapshot, suggestion); err != nil {
-		return nil, err
-	} else if handled {
-		return handledResponse, nil
-	}
-	if decision, ok := s.routeUnanchoredSubmitInput(context.Background(), snapshot, suggestion, confirmRequired); ok {
-		if decision.Route == inputRouteSocialChat {
-			return s.socialChatInputResponse(decision), nil
-		}
-		suggestion = applyInputRouteDecision(suggestion, decision)
-	}
-	preferredDelivery, fallbackDelivery := deliveryPreferenceFromSubmit(params)
-	if !suggestion.RequiresConfirm {
-		preferredDelivery, fallbackDelivery = mergeSuggestedDeliveryPreference(preferredDelivery, fallbackDelivery, suggestion.DirectDeliveryType)
-	}
-	if s.intent.AnalyzeSnapshot(snapshot) == "waiting_input" {
-		task := s.runEngine.CreateTask(runengine.CreateTaskInput{
-			SessionID:         stringValue(params, "session_id", ""),
-			RequestSource:     stringValue(params, "source", ""),
-			RequestTrigger:    stringValue(params, "trigger", ""),
-			Title:             "等待补充输入",
-			SourceType:        suggestion.TaskSourceType,
-			Status:            "waiting_input",
-			Intent:            nil,
-			PreferredDelivery: preferredDelivery,
-			FallbackDelivery:  fallbackDelivery,
-			CurrentStep:       "collect_input",
-			RiskLevel:         s.risk.DefaultLevel(),
-			Timeline:          initialTimeline("waiting_input", "collect_input"),
-			Snapshot:          snapshot,
-		})
-
-		bubble := s.delivery.BuildBubbleMessage(task.TaskID, "status", "请先告诉我你希望我处理什么内容。", task.StartedAt.Format(dateTimeLayout))
-		if _, ok := s.runEngine.SetPresentation(task.TaskID, bubble, nil, nil); ok {
-			task, _ = s.runEngine.GetTask(task.TaskID)
-		}
-
-		return map[string]any{
-			"task":            taskMap(task),
-			"bubble_message":  bubble,
-			"delivery_result": nil,
-		}, nil
-	}
-
-	task := s.runEngine.CreateTask(runengine.CreateTaskInput{
-		SessionID:         stringValue(params, "session_id", ""),
-		RequestSource:     stringValue(params, "source", ""),
-		RequestTrigger:    stringValue(params, "trigger", ""),
-		Title:             suggestion.TaskTitle,
-		SourceType:        suggestion.TaskSourceType,
-		Status:            taskStatusForSuggestion(suggestion.RequiresConfirm),
-		Intent:            suggestion.Intent,
-		PreferredDelivery: preferredDelivery,
-		FallbackDelivery:  fallbackDelivery,
-		CurrentStep:       currentStepForSuggestion(suggestion.RequiresConfirm, suggestion.Intent),
-		RiskLevel:         s.risk.DefaultLevel(),
-		Timeline:          initialTimeline(taskStatusForSuggestion(suggestion.RequiresConfirm), currentStepForSuggestion(suggestion.RequiresConfirm, suggestion.Intent)),
-		Snapshot:          snapshot,
-	})
-	s.publishTaskStart(task.TaskID, task.SessionID, requestTraceID(params))
-	s.attachMemoryReadPlans(task.TaskID, task.RunID, snapshot, suggestion.Intent)
-
-	bubble := s.delivery.BuildBubbleMessage(task.TaskID, bubbleTypeForSuggestion(suggestion.RequiresConfirm), bubbleTextForInput(suggestion), task.StartedAt.Format(dateTimeLayout))
-	deliveryResult := map[string]any(nil)
-	if !suggestion.RequiresConfirm {
-		if queuedTask, queueBubble, queued, queueErr := s.queueTaskIfSessionBusy(task); queueErr != nil {
-			return nil, queueErr
-		} else if queued {
-			task = queuedTask
-			bubble = queueBubble
-		} else {
-			governedTask, governedResponse, handled, governanceErr := s.handleTaskGovernanceDecision(task, suggestion.Intent)
-			if governanceErr != nil {
-				return nil, governanceErr
-			}
-			if handled {
-				return governedResponse, nil
-			}
-			task = governedTask
-			var execErr error
-			task, bubble, deliveryResult, _, execErr = s.executeTask(task, snapshot, suggestion.Intent)
-			if execErr != nil {
-				return nil, execErr
-			}
-		}
-	} else {
-		if _, ok := s.runEngine.SetPresentation(task.TaskID, bubble, nil, nil); ok {
-			task, _ = s.runEngine.GetTask(task.TaskID)
-		}
-	}
-
-	response := map[string]any{
-		"task":            taskMap(task),
-		"bubble_message":  bubble,
-		"delivery_result": nil,
-	}
-	if deliveryResult != nil {
-		response["delivery_result"] = deliveryResult
-	}
-
-	return response, nil
-}
-
-// StartTask handles agent.task.start and creates the task/run mapping from an
-// explicit or inferred intent.
-func (s *Service) StartTask(params map[string]any) (map[string]any, error) {
-	snapshot := s.context.Capture(params)
-	explicitIntent := mapValue(params, "intent")
-	options := mapValue(params, "options")
-	forceConfirmRequired := boolValue(options, "confirm_required", false)
-	confirmRequired := taskStartConfirmRequired(snapshot, explicitIntent, forceConfirmRequired)
-	if response, handled, resolvedSessionID, err := s.maybeContinueExistingTask(params, snapshot, explicitIntent, taskContinuationOptions{
-		ConfirmRequired:      confirmRequired,
-		ForceConfirmRequired: forceConfirmRequired,
-	}); err != nil {
-		return nil, err
-	} else if handled {
-		return response, nil
-	} else if strings.TrimSpace(resolvedSessionID) != "" {
-		params = withResolvedSessionID(params, resolvedSessionID)
-	}
-	if handledResponse, handled, err := s.handleScreenAnalyzeStart(params, snapshot, explicitIntent); err != nil {
-		return nil, err
-	} else if handled {
-		return handledResponse, nil
-	}
-	suggestion := s.intent.Suggest(snapshot, explicitIntent, confirmRequired)
-	fallbackConfirmRequired := confirmRequired
-	// Screen inference already carries its own authorization boundary; only an
-	// explicit caller request should turn an unavailable screen path back into
-	// intent confirmation.
-	if stringValue(suggestion.Intent, "name", "") == "screen_analyze" && !forceConfirmRequired {
-		fallbackConfirmRequired = suggestion.RequiresConfirm
-	}
-	suggestion = s.normalizeSuggestedIntentForAvailability(snapshot, suggestion, fallbackConfirmRequired)
-	if handledResponse, handled, err := s.handleScreenAnalyzeSuggestion(params, snapshot, suggestion); err != nil {
-		return nil, err
-	} else if handled {
-		return handledResponse, nil
-	}
-	preferredDelivery, fallbackDelivery := deliveryPreferenceFromStart(params)
-	if len(explicitIntent) == 0 && !suggestion.RequiresConfirm {
-		preferredDelivery, fallbackDelivery = mergeSuggestedDeliveryPreference(preferredDelivery, fallbackDelivery, suggestion.DirectDeliveryType)
-	}
-
-	task := s.runEngine.CreateTask(runengine.CreateTaskInput{
-		SessionID:         stringValue(params, "session_id", ""),
-		RequestSource:     stringValue(params, "source", ""),
-		RequestTrigger:    stringValue(params, "trigger", ""),
-		Title:             suggestion.TaskTitle,
-		SourceType:        suggestion.TaskSourceType,
-		Status:            taskStatusForSuggestion(suggestion.RequiresConfirm),
-		Intent:            suggestion.Intent,
-		PreferredDelivery: preferredDelivery,
-		FallbackDelivery:  fallbackDelivery,
-		CurrentStep:       currentStepForSuggestion(suggestion.RequiresConfirm, suggestion.Intent),
-		RiskLevel:         s.risk.DefaultLevel(),
-		Timeline:          initialTimeline(taskStatusForSuggestion(suggestion.RequiresConfirm), currentStepForSuggestion(suggestion.RequiresConfirm, suggestion.Intent)),
-		Snapshot:          snapshot,
-	})
-	s.publishTaskStart(task.TaskID, task.SessionID, requestTraceID(params))
-	s.attachMemoryReadPlans(task.TaskID, task.RunID, snapshot, suggestion.Intent)
-
-	bubble := s.delivery.BuildBubbleMessage(task.TaskID, bubbleTypeForSuggestion(suggestion.RequiresConfirm), bubbleTextForStart(suggestion), task.StartedAt.Format(dateTimeLayout))
-	response := map[string]any{
-		"task":            taskMap(task),
-		"bubble_message":  bubble,
-		"delivery_result": nil,
-	}
-
-	if suggestion.RequiresConfirm {
-		if _, ok := s.runEngine.SetPresentation(task.TaskID, bubble, nil, nil); ok {
-			task, _ = s.runEngine.GetTask(task.TaskID)
-			response["task"] = taskMap(task)
-		}
-		return response, nil
-	}
-
-	if queuedTask, queueBubble, queued, queueErr := s.queueTaskIfSessionBusy(task); queueErr != nil {
-		return nil, queueErr
-	} else if queued {
-		response["task"] = taskMap(queuedTask)
-		response["bubble_message"] = queueBubble
-		return response, nil
-	}
-
-	governedTask, governedResponse, handled, governanceErr := s.handleTaskGovernanceDecision(task, suggestion.Intent)
-	if governanceErr != nil {
-		return nil, governanceErr
-	}
-	if handled {
-		return governedResponse, nil
-	}
-	task = governedTask
-
-	deliveryResult := map[string]any(nil)
-	var execErr error
-	task, bubble, deliveryResult, _, execErr = s.executeTask(task, snapshot, suggestion.Intent)
-	if execErr != nil {
-		return nil, execErr
-	}
-	response["task"] = taskMap(task)
-	response["bubble_message"] = bubble
-	if len(deliveryResult) > 0 {
-		response["delivery_result"] = deliveryResult
-	} else {
-		response["delivery_result"] = nil
-	}
-	return response, nil
-}
-
-// taskStartConfirmRequired keeps confirmation as an explicit pre-execution gate.
-// Object-based task starts with their own instruction can enter the Agent Loop
-// directly, while bare objects still stop for intent confirmation.
-func taskStartConfirmRequired(snapshot contextsvc.TaskContextSnapshot, explicitIntent map[string]any, forceConfirm bool) bool {
-	if forceConfirm {
-		return true
-	}
-	if len(explicitIntent) > 0 {
-		return false
-	}
-	return !taskStartHasExplicitGoal(snapshot)
-}
-
-func taskStartHasExplicitGoal(snapshot contextsvc.TaskContextSnapshot) bool {
-	switch snapshot.InputType {
-	case "file":
-		return strings.TrimSpace(snapshot.Text) != ""
-	default:
-		return false
-	}
-}
-
 func (s *Service) handleScreenAnalyzeStart(params map[string]any, snapshot contextsvc.TaskContextSnapshot, explicitIntent map[string]any) (map[string]any, bool, error) {
 	if stringValue(explicitIntent, "name", "") != "screen_analyze" || s.executor == nil || !s.executor.ScreenCapabilitySnapshot().Available {
 		return nil, false, nil
@@ -890,105 +638,6 @@ func (s *Service) activeApprovalIDForTask(task runengine.TaskRecord) (string, bo
 		return "", false
 	}
 	return approvalID, true
-}
-
-// ConfirmTask handles agent.task.confirm.
-// It only accepts tasks that are still waiting in the intent confirmation phase,
-// then either keeps clarification open, applies a corrected intent, or confirms
-// the stored intent before continuing through governance and delivery.
-func (s *Service) ConfirmTask(params map[string]any) (map[string]any, error) {
-	taskID := stringValue(params, "task_id", "")
-	task, ok := s.runEngine.GetTask(taskID)
-	if !ok {
-		return nil, ErrTaskNotFound
-	}
-	if task.Status != "confirming_intent" {
-		return nil, ErrTaskStatusInvalid
-	}
-	confirmed := boolValue(params, "confirmed", false)
-	correctedIntent := mapValue(params, "corrected_intent")
-	intentValue := cloneMap(task.Intent)
-	if !confirmed && len(correctedIntent) > 0 {
-		intentValue = correctedIntent
-	}
-	if !confirmed && len(correctedIntent) == 0 {
-		updatedTask, err := s.revertTaskToIntentConfirmation(task)
-		if err != nil {
-			return nil, err
-		}
-		bubble := s.delivery.BuildBubbleMessage(task.TaskID, "status", "这不是我该做的处理方式。请重新说明你的目标，或给我一个更准确的处理意图。", updatedTask.UpdatedAt.Format(dateTimeLayout))
-		if presentedTask, ok := s.runEngine.SetPresentation(task.TaskID, bubble, nil, nil); ok {
-			updatedTask = presentedTask
-		} else {
-			return nil, ErrTaskNotFound
-		}
-		return map[string]any{
-			"task":            taskMap(updatedTask),
-			"bubble_message":  bubble,
-			"delivery_result": nil,
-		}, nil
-	}
-	if strings.TrimSpace(stringValue(intentValue, "name", "")) == "" {
-		bubble := s.delivery.BuildBubbleMessage(task.TaskID, "status", "请先明确告诉我你希望执行的处理方式。", task.UpdatedAt.Format(dateTimeLayout))
-		if updatedTask, ok := s.runEngine.SetPresentation(task.TaskID, bubble, nil, nil); ok {
-			return map[string]any{
-				"task":            taskMap(updatedTask),
-				"bubble_message":  bubble,
-				"delivery_result": nil,
-			}, nil
-		}
-		return nil, ErrTaskNotFound
-	}
-	updatedTitle := s.intent.Suggest(snapshotFromTask(task), intentValue, false).TaskTitle
-
-	bubble := s.delivery.BuildBubbleMessage(task.TaskID, "status", "已按新的要求开始处理", task.UpdatedAt.Format(dateTimeLayout))
-	updatedTask, ok := s.runEngine.UpdateIntent(task.TaskID, updatedTitle, intentValue)
-	if !ok {
-		return nil, ErrTaskNotFound
-	}
-	s.attachMemoryReadPlans(updatedTask.TaskID, updatedTask.RunID, snapshotFromTask(updatedTask), intentValue)
-	if queuedTask, queueBubble, queued, queueErr := s.queueTaskIfSessionBusy(updatedTask); queueErr != nil {
-		return nil, queueErr
-	} else if queued {
-		return map[string]any{
-			"task":            taskMap(queuedTask),
-			"bubble_message":  queueBubble,
-			"delivery_result": nil,
-		}, nil
-	}
-	governedTask, governedResponse, handled, governanceErr := s.handleTaskGovernanceDecision(updatedTask, intentValue)
-	if governanceErr != nil {
-		return nil, governanceErr
-	}
-	if handled {
-		return governedResponse, nil
-	}
-	updatedTask = governedTask
-
-	updatedTask, ok = s.runEngine.ConfirmTask(task.TaskID, updatedTitle, intentValue, bubble)
-	if !ok {
-		return nil, ErrTaskNotFound
-	}
-	snapshot := snapshotFromTask(updatedTask)
-
-	updatedTask, resultBubble, deliveryResult, _, err := s.executeTask(updatedTask, snapshot, intentValue)
-	if err != nil {
-		return nil, err
-	}
-
-	return map[string]any{
-		"task":            taskMap(updatedTask),
-		"bubble_message":  resultBubble,
-		"delivery_result": optionalFormalDeliveryResult(deliveryResult),
-	}, nil
-}
-
-func (s *Service) revertTaskToIntentConfirmation(task runengine.TaskRecord) (runengine.TaskRecord, error) {
-	updatedTask, ok := s.runEngine.UpdateIntent(task.TaskID, confirmationTitleFromTask(task), nil)
-	if !ok {
-		return runengine.TaskRecord{}, ErrTaskNotFound
-	}
-	return updatedTask, nil
 }
 
 // RecommendationGet handles agent.recommendation.get and returns lightweight
@@ -1575,53 +1224,6 @@ func parseEventTimeFilter(value string) time.Time {
 	return time.Time{}
 }
 
-// TaskSteer handles agent.task.steer by persisting one follow-up instruction for
-// a still-active task so later execution or resume paths can consume it.
-func (s *Service) TaskSteer(params map[string]any) (map[string]any, error) {
-	taskID := stringValue(params, "task_id", "")
-	message := stringValue(params, "message", "")
-	if strings.TrimSpace(taskID) == "" {
-		return nil, errors.New("task_id is required")
-	}
-	if strings.TrimSpace(message) == "" {
-		return nil, errors.New("message is required")
-	}
-	task, ok := s.runEngine.GetTask(taskID)
-	if !ok {
-		return nil, ErrTaskNotFound
-	}
-	if !taskCanAcceptExplicitSteering(task) {
-		return nil, ErrTaskStatusInvalid
-	}
-	bubble := s.delivery.BuildBubbleMessage(task.TaskID, "status", "已记录新的补充要求，后续执行会纳入该指令。", time.Now().Format(dateTimeLayout))
-	updatedTask, changed := s.runEngine.AppendSteeringMessage(task.TaskID, message, bubble)
-	if !changed {
-		return nil, ErrTaskStatusInvalid
-	}
-	return map[string]any{
-		"task":           taskMap(updatedTask),
-		"bubble_message": bubble,
-	}, nil
-}
-
-func taskCanAcceptExplicitSteering(task runengine.TaskRecord) bool {
-	switch task.Status {
-	case "processing":
-		// Active processing tasks can only consume steering when the running
-		// agent loop polls between rounds. Other processing paths must finish or
-		// queue a separate task instead of pretending the guidance was consumed.
-		return taskCanConsumeActiveSteering(task)
-	case "waiting_auth", "blocked":
-		// Deferred execution paths can carry explicit steering until approval or
-		// queue release resumes the task. Pending-input states are intentionally
-		// rejected so callers re-enter agent.input.submit and merge the text into
-		// the formal continuation snapshot instead of hiding it in runtime notes.
-		return true
-	default:
-		return false
-	}
-}
-
 // TaskArtifactList handles `agent.task.artifact.list` and returns protocol-ready
 // artifact items.
 func (s *Service) TaskArtifactList(params map[string]any) (map[string]any, error) {
@@ -1878,134 +1480,6 @@ func normalizeTaskDetailDeliveryResult(taskID string, deliveryResult map[string]
 		return nil
 	}
 	return normalizeDeliveryOpenResult(nil, cloneMap(deliveryResult), taskID)
-}
-
-// TaskControl handles agent.task.control and converts user actions into runtime
-// state-machine transitions. The orchestration layer owns error translation and
-// post-transition follow-up such as human-loop resume handling and queue drain,
-// because those behaviors depend on task-centric semantics rather than the raw
-// runtime mutation alone.
-func (s *Service) TaskControl(params map[string]any) (map[string]any, error) {
-	taskID := stringValue(params, "task_id", "")
-	if strings.TrimSpace(taskID) == "" {
-		return nil, errors.New("task_id is required")
-	}
-	action := stringValue(params, "action", "")
-	if strings.TrimSpace(action) == "" {
-		return nil, errors.New("action is required")
-	}
-	if !isSupportedTaskControlAction(action) {
-		return nil, fmt.Errorf("unsupported task control action: %s", action)
-	}
-	previousTask := runengine.TaskRecord{}
-	if existingTask, ok := s.runEngine.GetTask(taskID); ok {
-		previousTask = existingTask
-	}
-	wasHumanLoop := false
-	var reviewDecision map[string]any
-	arguments := mapValue(params, "arguments")
-	if action == "resume" {
-		wasHumanLoop = taskIsBlockedHumanLoop(previousTask)
-		if wasHumanLoop {
-			decision, decisionErr := humanReviewDecisionFromParams(arguments)
-			if decisionErr != nil {
-				return nil, decisionErr
-			}
-			reviewDecision = decision
-		}
-	}
-	bubble := s.delivery.BuildBubbleMessage(taskID, "status", controlBubbleText(action), currentTimeFromTask(s.runEngine, taskID))
-	updatedTask := runengine.TaskRecord{}
-	if action == "restart" {
-		preRestartTask, preparedRestartTask, restartErr := s.runEngine.PrepareRestart(taskID, bubble)
-		if restartErr != nil {
-			switch {
-			case errors.Is(restartErr, runengine.ErrTaskNotFound):
-				return nil, ErrTaskNotFound
-			case errors.Is(restartErr, runengine.ErrTaskStatusInvalid):
-				return nil, ErrTaskStatusInvalid
-			case errors.Is(restartErr, runengine.ErrTaskAlreadyFinished):
-				return nil, ErrTaskAlreadyFinished
-			default:
-				return nil, restartErr
-			}
-		}
-		previousTask = preRestartTask
-		updatedTask = preparedRestartTask
-	} else {
-		nextTask, err := s.runEngine.ControlTask(taskID, action, bubble)
-		if err != nil {
-			switch {
-			case errors.Is(err, runengine.ErrTaskNotFound):
-				return nil, ErrTaskNotFound
-			case errors.Is(err, runengine.ErrTaskStatusInvalid):
-				return nil, ErrTaskStatusInvalid
-			case errors.Is(err, runengine.ErrTaskAlreadyFinished):
-				return nil, ErrTaskAlreadyFinished
-			default:
-				return nil, err
-			}
-		}
-		updatedTask = nextTask
-	}
-	if action == "resume" && wasHumanLoop {
-		if traceResumedTask, traceBubble, _, resumed, resumeErr := s.resumeHumanLoopTask(updatedTask, reviewDecision); resumeErr != nil {
-			return nil, resumeErr
-		} else if resumed {
-			updatedTask = traceResumedTask
-			bubble = traceBubble
-		}
-	}
-	if action == "restart" {
-		restartedTask, restartBubble, restartErr := s.advanceRestartedTaskAttempt(previousTask, updatedTask)
-		if restartErr != nil {
-			return nil, restartErr
-		}
-		updatedTask = restartedTask
-		if restartBubble != nil {
-			bubble = restartBubble
-		}
-	}
-	if taskIsTerminal(updatedTask.Status) {
-		if queueErr := s.drainSessionQueue(updatedTask.SessionID); queueErr != nil {
-			return nil, queueErr
-		}
-	}
-
-	return map[string]any{
-		"task":           taskMap(updatedTask),
-		"bubble_message": bubble,
-	}, nil
-}
-
-// advanceRestartedTaskAttempt sends a fresh restart run through the same
-// pre-execution gates as a new task. Restart may allocate a new run_id, but it
-// must not bypass session serialization or the authorization boundary before
-// the executor receives that run.
-func (s *Service) advanceRestartedTaskAttempt(previousTask, task runengine.TaskRecord) (runengine.TaskRecord, map[string]any, error) {
-	if queuedTask, queueBubble, queued, queueErr := s.queueTaskIfSessionBusy(task); queueErr != nil {
-		return runengine.TaskRecord{}, nil, queueErr
-	} else if queued {
-		return queuedTask, queueBubble, nil
-	}
-
-	governedTask, governedResponse, handled, governanceErr := s.handleTaskGovernanceDecision(task, task.Intent)
-	if governanceErr != nil {
-		return runengine.TaskRecord{}, nil, governanceErr
-	}
-	if handled {
-		bubble := mapValue(governedResponse, "bubble_message")
-		if len(bubble) == 0 {
-			bubble = governedTask.BubbleMessage
-		}
-		return governedTask, bubble, nil
-	}
-
-	restartedTask, restartBubble, _, _, restartErr := s.executeTaskAttempt(previousTask, governedTask, snapshotFromTask(governedTask), governedTask.Intent)
-	if restartErr != nil {
-		return runengine.TaskRecord{}, nil, restartErr
-	}
-	return restartedTask, restartBubble, nil
 }
 
 // TaskInspectorConfigGet handles agent.task_inspector.config.get.
@@ -3245,90 +2719,6 @@ func taskMap(record runengine.TaskRecord) map[string]any {
 		result["finished_at"] = record.FinishedAt.Format(dateTimeLayout)
 	}
 	return result
-}
-
-func taskSessionValue(sessionID string) any {
-	if strings.TrimSpace(sessionID) == "" {
-		return nil
-	}
-	return strings.TrimSpace(sessionID)
-}
-
-func (s *Service) queueTaskIfSessionBusy(task runengine.TaskRecord) (runengine.TaskRecord, map[string]any, bool, error) {
-	activeTask, ok := s.runEngine.ActiveSessionTask(task.SessionID, task.TaskID)
-	if !ok {
-		return runengine.TaskRecord{}, nil, false, nil
-	}
-
-	bubble := s.delivery.BuildBubbleMessage(
-		task.TaskID,
-		"status",
-		fmt.Sprintf("当前会话已有任务 %s 正在执行，本任务已排队等待。", truncateText(activeTask.Title, 24)),
-		task.UpdatedAt.Format(dateTimeLayout),
-	)
-	var queuedTask runengine.TaskRecord
-	changed := false
-	if s.isPreparedRestartAttempt(task) {
-		queuedTask, changed = s.runEngine.QueuePreparedTaskForSession(task, activeTask.TaskID, bubble)
-	} else {
-		queuedTask, changed = s.runEngine.QueueTaskForSession(task.TaskID, activeTask.TaskID, bubble)
-	}
-	if !changed {
-		return runengine.TaskRecord{}, nil, false, ErrTaskNotFound
-	}
-	return queuedTask, bubble, true, nil
-}
-
-func (s *Service) drainSessionQueue(sessionID string) error {
-	for {
-		nextTask, ok := s.runEngine.NextQueuedTaskForSession(sessionID)
-		if !ok {
-			return nil
-		}
-		if activeTask, busy := s.runEngine.ActiveSessionTask(sessionID, nextTask.TaskID); busy && activeTask.TaskID != "" {
-			return nil
-		}
-
-		bubble := s.delivery.BuildBubbleMessage(
-			nextTask.TaskID,
-			"status",
-			"前序任务已完成，当前会话中的下一个任务开始执行。",
-			nextTask.UpdatedAt.Format(dateTimeLayout),
-		)
-		resumedTask, changed := s.runEngine.ResumeQueuedTask(nextTask.TaskID, s.activeExecutionStepName(nextTask.Intent), bubble)
-		if !changed {
-			return ErrTaskNotFound
-		}
-		resumedTask, handled, controlledErr := s.resumeQueuedControlledTask(resumedTask)
-		if controlledErr != nil {
-			return controlledErr
-		}
-		if handled {
-			if taskIsTerminal(resumedTask.Status) {
-				continue
-			}
-			return nil
-		}
-
-		governedTask, _, handled, governanceErr := s.handleTaskGovernanceDecision(resumedTask, resumedTask.Intent)
-		if governanceErr != nil {
-			return governanceErr
-		}
-		if handled {
-			if taskIsTerminal(governedTask.Status) {
-				continue
-			}
-			return nil
-		}
-
-		updatedTask, _, _, _, err := s.executeTask(governedTask, snapshotFromTask(governedTask), governedTask.Intent)
-		if err != nil {
-			return err
-		}
-		if !taskIsTerminal(updatedTask.Status) {
-			return nil
-		}
-	}
 }
 
 func taskIsTerminal(status string) bool {
@@ -4999,124 +4389,6 @@ func (s *Service) taskTimelineFromStructuredStorage(taskID string) []runengine.T
 	return result
 }
 
-// taskStatusForSuggestion derives the initial task_status from the suggestion
-// confirmation requirement.
-func taskStatusForSuggestion(requiresConfirm bool) string {
-	if requiresConfirm {
-		return "confirming_intent"
-	}
-	return "processing"
-}
-
-// currentStepForSuggestion derives the initial current_step from the suggested
-// intent.
-func currentStepForSuggestion(requiresConfirm bool, taskIntent map[string]any) string {
-	if requiresConfirm {
-		return "intent_confirmation"
-	}
-	if stringValue(taskIntent, "name", "") == "agent_loop" {
-		return "agent_loop"
-	}
-	return "generate_output"
-}
-
-// bubbleTypeForSuggestion selects the outward-facing bubble type for the
-// suggestion result.
-func bubbleTypeForSuggestion(requiresConfirm bool) string {
-	if requiresConfirm {
-		return "intent_confirm"
-	}
-	return "result"
-}
-
-// bubbleTextForInput returns the bubble text for agent.input.submit flows.
-func bubbleTextForInput(suggestion intent.Suggestion) string {
-	if suggestion.RequiresConfirm {
-		if !suggestion.IntentConfirmed {
-			return "我还不确定你想如何处理这段内容，请确认目标。"
-		}
-		return confirmIntentText(suggestion.Intent)
-	}
-	return suggestion.ResultBubbleText
-}
-
-// bubbleTextForStart returns the bubble text for agent.task.start flows.
-func bubbleTextForStart(suggestion intent.Suggestion) string {
-	if suggestion.RequiresConfirm {
-		if !suggestion.IntentConfirmed {
-			return "我还不确定你想如何处理当前对象，请先确认。"
-		}
-		return confirmIntentText(suggestion.Intent)
-	}
-	return suggestion.ResultBubbleText
-}
-func confirmIntentText(taskIntent map[string]any) string {
-	switch stringValue(taskIntent, "name", "") {
-	case "translate":
-		return "你是想翻译这段内容吗？"
-	case "rewrite":
-		return "你是想改写这段内容吗？"
-	case "explain":
-		return "你是想解释这段内容吗？"
-	case "summarize":
-		return "你是想总结这段内容吗？"
-	case "write_file":
-		return "你是想把结果整理成文档吗？"
-	default:
-		return "请确认你希望我如何处理当前内容。"
-	}
-}
-
-// initialTimeline creates the first timeline step for a new task and derives
-// whether that step starts as pending or running.
-func initialTimeline(status, currentStep string) []runengine.TaskStepRecord {
-	stepStatus := "running"
-	if status == "confirming_intent" || status == "waiting_input" {
-		stepStatus = "pending"
-	}
-
-	outputSummary := "等待继续处理"
-	if status == "waiting_input" {
-		outputSummary = "等待用户补充输入"
-	}
-
-	return []runengine.TaskStepRecord{
-		{
-			StepID:        fmt.Sprintf("step_%s", currentStep),
-			Name:          currentStep,
-			Status:        stepStatus,
-			OrderIndex:    1,
-			InputSummary:  "已识别到当前任务对象",
-			OutputSummary: outputSummary,
-		},
-	}
-}
-
-// controlBubbleText returns the status bubble text for a task_control action.
-func controlBubbleText(action string) string {
-	switch action {
-	case "pause":
-		return "任务已暂停"
-	case "resume":
-		return "任务已继续执行"
-	case "cancel":
-		return "任务已取消"
-	case "restart":
-		return "任务已重新开始"
-	default:
-		return "任务状态已更新"
-	}
-}
-
-func isSupportedTaskControlAction(action string) bool {
-	switch action {
-	case "pause", "resume", "cancel", "restart":
-		return true
-	default:
-		return false
-	}
-}
-
 // defaultIntentMap creates a minimal default intent payload for notepad
 // conversions.
 func defaultIntentMap(name string) map[string]any {
@@ -6407,20 +5679,6 @@ func deliveryTypeFromIntent(taskIntent map[string]any) string {
 	default:
 		return "workspace_document"
 	}
-}
-
-// deliveryPreferenceFromSubmit reads delivery preferences from
-// agent.input.submit. Submit uses options.* while agent.task.start uses a
-// dedicated delivery object, so the orchestrator keeps both decoders separate
-// and normalizes them before any execution or approval plan is built.
-func deliveryPreferenceFromSubmit(params map[string]any) (string, string) {
-	options := mapValue(params, "options")
-	return stringValue(options, "preferred_delivery", ""), ""
-}
-
-func deliveryPreferenceFromStart(params map[string]any) (string, string) {
-	deliveryOptions := mapValue(params, "delivery")
-	return stringValue(deliveryOptions, "preferred", ""), stringValue(deliveryOptions, "fallback", "")
 }
 
 func (s *Service) applyGovernanceAssessment(plan map[string]any, assessment execution.GovernanceAssessment) map[string]any {
